requirements.txt
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-v5==1.0.4
django-bootstrap4==3.0.0
django-js-asset==1.2.2
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
Pillow==8.1.2
gunicorn==20.1.0
sphinx
mkdocs
coverage
